careconnect_server/requirements.txt
# careconnect_server/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-generativeai
python-dotenv
websockets